vet-edit-report - move a whole bunch of things into vet/_lib
