+++ src/app/vet/(logged-in)/schedule/page.tsx
@@ -2,7 +2,7 @@ import { getAuthenticatedVetActor } from "@/lib/auth";
 import { RoutePath } from "@/lib/route-path";
 import { getAvailableDogs } from "@/lib/vet/actions/get-available-dogs";
 import { redirect } from "next/navigation";
-import { AppointmentScheduler } from "./_components/appointment-scheduler";
+import { AppointmentScheduler } from "../../_lib/components/appointment-scheduler";
 
 export default async function Page() {
   const actor = await getAuthenticatedVetActor();
