+++ src/app/vet/_lib/components/scheduler-badges.tsx
