+++ src/app/vet/_lib/models/scheduler-outcome.ts
@@ -1,0 +1,5 @@
+import { CALL_OUTCOME } from "@/lib/data/db-enums";
+
+export type SchedulerOutcome =
+  | typeof CALL_OUTCOME.APPOINTMENT
+  | typeof CALL_OUTCOME.DECLINED;
